Authentication added

+++ src/blog/views.py
@@ -1,8 +1,9 @@
 from django.shortcuts import render, get_object_or_404, redirect
-from django.contrib.auth.decorators import login_required
+from django.contrib.auth.decorators import login_required, permission_required
 from django.contrib.admin.views.decorators import staff_member_required
 from .models import BlogPost
-from django.http import Http404
+from django.contrib.auth.forms import UserCreationForm
+from django.contrib.auth import authenticate, login, logout
 from .forms import BlogPostForm, BlogPostModelForm
 
 
@@ -18,18 +19,22 @@ def blog_post_list_view(request):
     return render(request, 'blog/list.html', context)
 
 
-# @login_required
-@staff_member_required
+# @staff_member_required
 def blog_post_create_view(request):
-    form = BlogPostModelForm(request.POST or None, request.FILES or None)
-    if form.is_valid():
-        obj = form.save(commit=False)
-        obj.user = request.user
-        obj.save()
-        form = BlogPostModelForm()
+    if request.user.is_active:
+        form = BlogPostModelForm(request.POST or None, request.FILES or None)
+        if form.is_valid():
+            obj = form.save(commit=False)
+            obj.user = request.user
+            obj.save()
+            form = BlogPostModelForm()
+            return redirect(blog_post_detail_view, slug=obj.slug)
+    else:
+        return redirect('/accounts/login/')
 
     context = {
-        'form': form
+        'form': form,
+        'title': "Create your blog here"
     }
     return render(request, 'form.html', context)
 
@@ -43,12 +48,15 @@ def blog_post_detail_view(request, slug):
     return render(request, 'blog/detail.html', context)
 
 
-@staff_member_required
+#@staff_member_required
+#@permission_required('BlogPost.can_publish_blog')
 def blog_post_update_view(request, slug):
     obj = get_object_or_404(BlogPost, slug=slug)
     form = BlogPostModelForm(request.POST or None, instance=obj)
     if form.is_valid():
         form.save()
+        sl = form.cleaned_data["slug"]
+        return redirect(blog_post_detail_view,slug=sl)
     context = {
         'title': f"Update {obj.title}",
         'form': form,
@@ -56,13 +64,41 @@ def blog_post_update_view(request, slug):
     return render(request, 'form.html', context)
 
 
-@staff_member_required
+# @staff_member_required
+#@permission_required('BlogPost.can_publish_blog')
 def blog_post_delete_view(request, slug):
-    obj = get_object_or_404(BlogPost, slug=slug)
-    if request.method == "POST":
-        obj.delete()
-        return redirect("/blog")
+    if request.user.is_active:
+        obj = get_object_or_404(BlogPost, slug=slug)
+        if request.method == "POST":
+            obj.delete()
+            return redirect("/blog")
+    else:
+        return redirect('/accounts/login/')
+
     context = {
         'object': obj
     }
     return render(request, 'blog/delete.html', context)
+
+
+def register(request):
+    if request.method =="POST":
+        form = UserCreationForm(request.POST)
+
+        if form.is_valid():
+            form.save()
+            username = form.cleaned_data['username']
+            password = form.cleaned_data['password1']
+            user = authenticate(username=username, password=password)
+            login(request, user)
+            return redirect('/blog')
+    else:
+        form = UserCreationForm()
+    context={
+        'form': form
+    }
+    return render(request, 'registration/register.html', context)
+
+def logout_view(request):
+    logout(request)
+    return redirect("/")
